requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
celery
redis
django-celery-beat
django-celery-results
Pillow
requests
python-jose[cryptography]
django-environ
gunicorn
whitenoise
django-allauth
django-rest-auth
djangorestframework-simplejwt
